Update CPP calculation for low periodic income
The CPP calculation presented in the T4127 is somewhat vague on low
periodic income and high bonuses. Instead of using PI, I has to be used
for basic CPP calculation and a flat rate of 5.95% on B.

Some changes might be required for the implementer interface.

Reference: https://gethumi.atlassian.net/browse/PAY-427