fix: Always use MAX CPP/EI amounts when reached
BUG-8456 showed that our approach to calculating the K2 and K2P values
did not properly take into account when the contribution maximum had
been reached.

This commit implements a new calculation method that Greg found to work
in his tax calculating spreadsheet, but is now diviating from the
written formulas in T4127#117th and T4127#118th edition. It however
matches PDOC.

One side effect is a slight value change to the calculated taxes on
payrolls with no periodic income, but with non-periodic income.