Add rounding to bpans
The CRA text is hilariously bad:

"If Nova Scotia’s basic personal amount (BPANS) has three or more digits
after the decimal point, increase the second digit by one if the third
digit is five or more, and drop the third digit. If the third digit
after the decimal point is less than five, drop the third digit."

So, round to the cent (which we use as the base value).