Update release-please to v4
The new version needs a dedicated manifest file to set the
configurations that we had been setting before in v3. This has not
really been tested, but should be triggered by PRs merged into master
and will then let us know if things work as expected.

Details on the v3 to v4 upgrade:
https://github.com/google-github-actions/release-please-action#upgrading-from-v3-to-v4